+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# NEXT UP
+
+- Form of the beast effect
+
 # 3.4.17
 
 - Some adventure muncher scenes with tiles would not import the tile, if the tile config was generated in v10 of Foundry.